refactor motor crusade is coming
